Add solution, change gulp to parcel, use MSBuild to create NuGet package and UmbPack for Umbraco package
