+++ .gitignore
@@ -1,6 +1,6 @@
+.vs/
+bin/
+obj/
 node_modules/
-dist/
-buildpackage/artifacts/*.zip
-buildpackage/artifacts/*.nupkg
-/.vs/NestingContently/v16
-.vs/VSWorkspaceState.json
+*.user
+*.min.js
