+++ NestingContently.Umbraco/App_Plugins/NestingContently/backoffice/styles.css
@@ -1,0 +1,4 @@
+.umb-block-list__block--disabled,
+.umb-nested-content__item--disabled {
+	opacity: 0.5;
+}
