+++ NestingContently.Umbraco/package.json
@@ -1,0 +1,22 @@
+{
+	"name": "nestingcontently.umbraco",
+	"version": "1.0.0",
+	"private": true,
+	"scripts": {
+		"dev": "parcel build src/entry.js --out-dir App_Plugins/NestingContently/backoffice --out-file nesting-contently.min.js --no-source-maps --no-minify --no-cache",
+		"build": "parcel build src/entry.js --out-dir App_Plugins/NestingContently/backoffice --out-file nesting-contently.min.js --no-source-maps --no-cache"
+	},
+	"babel": {
+		"presets": [
+			"@babel/env"
+		]
+	},
+	"browserslist": [
+		"defaults"
+	],
+	"devDependencies": {
+		"@babel/core": "^7.11.6",
+		"@babel/preset-env": "^7.11.5",
+		"parcel-bundler": "^1.12.4"
+	}
+}
