--- src/NestingContently/backoffice/editor.html
@@ -1,1 +1,0 @@
-<div ng-controller="nestingContentlyController"></div>
