--- src/NestingContently/backoffice/styles.css
@@ -1,4 +1,0 @@
-.umb-block-list__block--disabled,
-.umb-nested-content__item--disabled {
-    opacity: 0.5; 
-}
