--- src/NestingContently/package.prod.manifest
@@ -1,19 +1,0 @@
-﻿{  
-    propertyEditors: [      
-        {
-            alias: "NestingContently",
-            name: "Nesting Contently",
-            icon: "icon-bird",
-            editor: {
-                view: "~/app_plugins/nestingcontently/backoffice/editor.html",
-                valueType: "INT"
-            }
-        }        
-    ],
-    css: [
-        '~/app_plugins/nestingcontently/backoffice/styles.css'
-    ],
-    javascript: [	
-		      '~/app_plugins/nestingcontently/backoffice/nesting-contently.min.js'
-    ]
-}
